#299 fix: toast메세지 표시 위치 home으로 수정 및 둘러보기 클릭시 home으로 리다이렉트
