+++ src/app/page.tsx
@@ -3,11 +3,10 @@
 import { signIn } from "next-auth/react";
 import { Button } from "@/components/ui/button";
 import Image from "next/image";
-import { useRouter, useSearchParams } from "next/navigation";
+import { useRouter } from "next/navigation";
 import { useAuthRedirect } from "@/hooks/useAuthRedirect";
 // import LoadingPage from "@/components/common/LoadingPage";
-import { toast } from "sonner";
-import { useEffect, useState, useRef } from "react";
+import { useEffect, useState } from "react";
 
 import { motion, AnimatePresence } from "framer-motion";
 
@@ -19,51 +18,17 @@ const phrases = [
 
 export default function Home() {
   const router = useRouter();
-  const searchParams = useSearchParams();
   const { status, isLoading, isAuthenticated } = useAuthRedirect();
-  const [redirectPath, setRedirectPath] = useState("");
-  const hasShownToast = useRef(false);
 
   const [currentIndex, setCurrentIndex] = useState(0);
 
-  useEffect(() => {
-    const redirect = searchParams.get("redirect");
-
-    if (status === "unauthenticated" && redirect && !hasShownToast.current) {
-      hasShownToast.current = true;
-      setRedirectPath(redirect);
-
-      // Toaster가 완전히 마운트될 때까지 약간의 지연
-      setTimeout(() => {
-        try {
-          toast.warning("로그인이 필요한 페이지입니다.", {
-            // id: "login-required", // 고유 ID로 중복 방지
-            // action: {
-            //   label: "로그인하기",
-            //   onClick: () => handleLogin(),
-            // },
-            duration: 3000,
-          });
-        } catch (error) {
-          console.error("Toast 표시 실패:", error);
-        }
-      }, 100); // 100ms 지연
-
-      // URL에서 파라미터 제거
-      const newUrl = new URL(window.location.href);
-      newUrl.searchParams.delete("redirect");
-      window.history.replaceState({}, "", newUrl.toString());
-    }
-  }, [searchParams]); // searchParams 변경을 감지하도록 수정
-
   const handleLogin = () => {
     // 리다이렉트 경로가 있으면 로그인 후 해당 페이지로 이동
-    const callbackUrl = redirectPath || "/home";
-    signIn("kakao", { callbackUrl });
+    signIn("kakao", { callbackUrl: "/home" });
   };
 
   const handleExplore = () => {
-    router.push("/onboarding");
+    router.push("/home");
   };
 
   const tmp = () => {
